% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/business_logic_functions.R
\name{format_ship_type}
\alias{format_ship_type}
\title{To format the ship type to get the ship image file}
\usage{
format_ship_type(ship_type, suffix = "", file_extension = "png")
}
\arguments{
\item{ship_type}{A string with the ship's type}

\item{file_extension}{A string which contains file extension. Defaults to .png}

\item{type}{A string which contains the suffix that can be appended to the file name. Default is empty string}
}
\value{
A string which contains the file name corresponding to the ship type
}
\description{
To format the ship type to get the ship image file
}
\examples{
format_ship_type("Cargo")
}
