% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/business_logic_functions.R
\name{get_current_ship_data}
\alias{get_current_ship_data}
\title{To filter the current ship's data from the whole dataset}
\usage{
get_current_ship_data(ais_data, ship_id)
}
\arguments{
\item{ais_data}{A tbl_df/tbl/data.frame with the ais related ships data}

\item{ship_id}{A number with the ship's id}
}
\value{
A tibble with only this ship's data
}
\description{
To filter the current ship's data from the whole dataset
}
