% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_functions.R
\name{get_data_from_db}
\alias{get_data_from_db}
\title{To fetch data from database using select query}
\usage{
get_data_from_db(
  query,
  params = NULL,
  db_name = default_db_name,
  host_name = default_host_name,
  host_password = default_host_password,
  host_ip = default_host_ip
)
}
\arguments{
\item{query}{A string with the select query}

\item{params}{A list of optional parameters that fill the ? in the query}

\item{db_name}{A string which contains the name of the database}

\item{host_name}{A string which contains the username to access the database}

\item{host_password}{A string which contains the password to access the database}

\item{host_ip}{A string which contains the host address of the database}
}
\value{
A tibble which is the result of the query
}
\description{
To fetch data from database using select query
}
