% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/business_logic_functions.R
\name{get_longest_ais_gap_data}
\alias{get_longest_ais_gap_data}
\title{To filter the data with the longest distance measure between two consecutive AIS readings}
\usage{
get_longest_ais_gap_data(
  current_ship_data,
  geo_measure = "geodesic",
  round_accu = 0
)
}
\arguments{
\item{current_ship_data}{A tbl_df/tbl/data.frame with a particular ship's AIS data}

\item{geo_measure}{An optional string argument to set accuracy for the geo distance calculation. Can be "haversine" "vincenty", "geodesic", or "cheap"}

\item{round_accu}{An integer to specify the rounding digits accuracy}
}
\value{
A tibble with two consecutive rows of AIS data with the most distance travelled
}
\description{
To filter the data with the longest distance measure between two consecutive AIS readings
}
