% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/business_logic_functions.R
\name{get_ship_details}
\alias{get_ship_details}
\title{To filter the ship from the ships_data and return the basic ship details}
\usage{
get_ship_details(ships_data, ship_name)
}
\arguments{
\item{ships_data}{A dataframe containing the ships data required by this app}

\item{ship_name}{A string containing the ship name whos details has to be extracted}
}
\value{
A list with the details about the ship
}
\description{
To filter the ship from the ships_data and return the basic ship details
}
