% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_functions.R
\name{insert_into_db}
\alias{insert_into_db}
\title{To insert new values inside the MySQL table}
\usage{
insert_into_db(
  table_name,
  values,
  db_name = default_db_name,
  host_name = default_host_name,
  host_password = default_host_password,
  host_ip = default_host_ip
)
}
\arguments{
\item{table_name}{A string which contains the MySQL table name where the data has to be inserted}

\item{values}{A tbl_df/tbl/data.frame which is the same structure as the MySQL table}

\item{db_name}{A string which contains the name of the database}

\item{host_name}{A string which contains the username to access the database}

\item{host_password}{A string which contains the password to access the database}

\item{host_ip}{A string which contains the host address of the database}
}
\value{
Returns the result obtained after the insert, NULL if the data was inserted successfully
}
\description{
To insert new values inside the MySQL table
}
